fix layout of carousel
