+++ components/Custom/index.tsx
@@ -1,3 +1,4 @@
 export { default as Button } from './Button'
-export * from './FormComponents'
 export { default as Modal } from './Modal'
+export { default as Carousel } from './Carousel'
+export * from './FormComponents'
