+++ next.config.js
@@ -1,3 +1,7 @@
 module.exports = {
-  reactStrictMode: true
+  reactStrictMode: true,
+  // need to remove once test is deleted
+  images: {
+    domains: ['images.unsplash.com', 'tinyurl.com']
+  }
 }
